Nochmals Korrekturen, jetzt keine Warnings mehr! Motoren eingebunden

+++ INTRO_Common_TORO/Reflectance.c
@@ -159,16 +159,13 @@ static void REF_MeasureRaw(SensorTimeType raw[REF_NOF_SENSORS]) {
     timerVal = RefCnt_GetCounterValue(timerHandle);
     cnt = 0;
     for(i=0;i<REF_NOF_SENSORS;i++) {
-
-      if (raw[i]==MAX_SENSOR_VALUE)
-      { /* not measured yet? */
+      if (raw[i]==MAX_SENSOR_VALUE) { /* not measured yet? */
         if (SensorFctArray[i].GetVal()==0) {
-          raw[i] = timerVal;
+          raw[i] = (uint16_t)timerVal;
         }
-      } else
-	  { /* have value */
-		cnt++;
-	  }
+      } else { /* have value */
+        cnt++;
+      }
     }
   } while(cnt!=REF_NOF_SENSORS);
   //mutex bis hier
